feat: update profile data
